+++ __init__.py
@@ -3,6 +3,7 @@ from comfy_api.latest import ComfyExtension, io
 from .nodes import MiniMaxH3PromptEnhancer
 from .music3 import MiniMaxMusic3PromptEnhancer
 from .yue2 import YuE2MusicPromptEnhancer
+from .qwen_image21 import QwenImage21PromptEnhancer
 from .seedance20 import Seedance20PromptEnhancer
 from .case_library_routes import register_routes
 from .local_qwen_routes import register_local_qwen_routes
@@ -45,6 +46,7 @@ class T8PromptEnhancerExtension(ComfyExtension):
             T8CharacterPerformanceBibleStack,
             *CREATIVE_SUITE_NODES,
             YuE2MusicPromptEnhancer,
+            QwenImage21PromptEnhancer,
         ]
 
 
@@ -56,6 +58,7 @@ __all__ = [
     "WEB_DIRECTORY",
     "T8PromptEnhancerExtension",
     "MiniMaxMusic3PromptEnhancer",
+    "QwenImage21PromptEnhancer",
     "T8LLMProviderConfig",
     "T8PromptInspector",
     "T8PerformanceDirectorConfig",